Add files via upload
This is still a bit messy. I need to take out the output statements and the graph (unless we like the graph) showing the final alignment. Also, I'll need to add something that converts one of the input sequences to the reverse complement if the sequences are not both 5' -> 3', since the limitation on this code is that the input sequences need to be somewhat similar. Could just do a Needleman-Wunsch comparison of the first 100 bp of the sequences and see if they meet some similarity threshold, otherwise flip one to the reverse complement. Also, it's still in MATLAB. I wanted to convert it to R tonight, but there was an error that took longer to fix than anticipated (it's fixed now).

This is still basically a Needleman-Wunsch aligner, but it first searches both input sequences for regions of similarity (defined by input parameters, which I can adjust according to sequences) and replaces those regions with a number index. In the two CoVid sequences I tested, this reduced the size of the sequences from ~29000 to ~30, making it easily possible to align using NW. However, if the sequences are very dissimilar, the algorithm won't work, although I don't expect this to be the case.